Relax some of the constraints on using DeviceBuffer.
DeviceCopy types no longer have to implement Clone. Instead, functions that clone use that as an extra bound.

DeviceBox, DeviceBuffer and DeviceSlice can now contain non-DeviceCopy types when instantiated using `uninitialized()`.
This is useful for creating buffers of data which are initialized by a kernel as temporary scratch space and never copied to or from the device.