Fix NULL characters contained in Device::get_name()
When saving the output of Device::get_name() to a file, it becomes visible that
the device's name string is padded with NULL characters until the end of the array:

GeForce 940M^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@

The issue can be reproduced by saving the output of device::test::test_get_name
to a file and then viewing it with a text editor.

This fix filters out the NULL characters before converting the C string to a Rust
String.